+++ CHANGELOG.md
@@ -1,3 +1,40 @@
+
+##0.0.66
+
+* Fix bug for resume camera in android
+
+##0.0.65
+
+* Migrate from firebase-ml-vision to com.google.mlkit in Android
+
+##0.0.64
+
+* Update firebase-ml-vision in android
+
+##0.0.63
+
+* Improve pause camera CameraAPI 2
+
+##0.0.62
+
+* Improve pause and resume camera CameraAPI 2
+
+##0.0.61
+
+* Start Preview after take picture in CameraAPI 1
+
+##0.0.60
+
+* Improve reading barcode in Android
+
+##0.0.58
+
+* Fix bugs in orientation CameraAPI 1
+
+##0.0.57
+
+* Improve rotation in CameraView1
+
 ##0.0.54
 
 * Set taken picture size equal to preview size in CameraAPI 1
@@ -5,6 +42,7 @@
 ##0.0.53
 
 * Avoid take picture before starting preview session completely
+
 ##0.0.52
 
 * Orientation in CameraAPI 1 fixed